lesson-10-api-layer-ai-systems/demo/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
openai==1.54.3
